Events_groupbuy/requirements.txt
Django==2.2.3
uwsgi
psycopg2
django-crispy-forms
Pillow
django_social_share
stripe